Create parent marc mapper
